Create function that deletes a grid
This function allows the user to set the grid dimensions without
refreshing the page.

The while loop checks to see if the parent has any child nodes.
If it does, the child gets deleted and the loop runs until empty.

The function is called in the build grid to prevent the addition of
cells from a previously set grid.