Create a feature that fills and unfills a cell when clicked
Two functions where created: fillCellOnClick and createEventListeners.

The first simply checks to see if the cell is already black; if not, it
is filled black.

The second selects the divs with a class of 'cell' and creates event
listeners for each div with the 'cell' class. When the cell is clicked,
the fillCellOnClick function is called and changes the cell color based
on whether it is black or white.

createEventListeners is called at the end of buildGrid so it can be
looped through without checking if a grid has not been built yet.