Styling changes for cells, slider, and buttons
Cells:
Cells were changed to off-white giving the user a base color
that better suits most colors

Slider and buttons colors:
Sliders, default, and inactive buttons were given a lighter shade of
purple allowing the color to be distinctive from the surrounding
background. Adds importance to the buttons, but not while they are inactive
or needed.